增加根据注解设置分片键的功能

+++ src/main/java/com/qihoo/multids/sharding/ISharding.java
@@ -7,7 +7,7 @@ import java.util.List;
  */
 public interface ISharding {
 
-    String route(Long key);
+    String route(int key);
 
     void initNodes(List<String> nodes);
 }
